MCP_assesment2/requirements.txt
aiohttp
requests
Flask
python-dotenv
botbuilder-core
botbuilder-schema